Clean up test harness to use real React runtime exclusively
- Remove old component-runner.ts with mock implementations (SimpleRegistry, SimpleCompiler, etc.)
- Rename component-runner-v2.ts to component-runner.ts as it's now the only implementation
- Update all imports and exports to use the cleaned up naming
- Fix registry.components.keys() error by removing direct access to internal registry structure
- Remove all 'V2' references from class names and debug messages

The test harness now exclusively uses the real React runtime UMD bundle instead of mock implementations, providing better compatibility with the Angular implementation and reducing maintenance burden.
