+++ packages/React/test-harness/src/index.ts
@@ -1,7 +1,7 @@
 export { ReactTestHarness, TestHarnessOptions } from './lib/test-harness';
 export { BrowserManager, BrowserContextOptions } from './lib/browser-context';
+// Export the component runner that uses the real React runtime UMD bundle
 export { ComponentRunner, ComponentExecutionOptions, ComponentExecutionResult } from './lib/component-runner';
-// ComponentRunnerV2 is now used internally by default, no need to export separately
 export { AssertionHelpers } from './lib/assertion-helpers';
 export { ComponentLinter, LintResult, Violation, FixSuggestion } from './lib/component-linter';
 export { ComponentSpec } from '@memberjunction/interactive-component-types';
